requirements.txt
discord.py
python-dotenv
Flask
discord.py[voice]